backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database & Auth
supabase==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# NLP & ML (will install in Sprint 4 when needed)
# spacy==3.7.2  # Requires C++ compiler on Windows
scikit-learn==1.4.0
textstat==0.7.3
# numpy==1.26.3  # Installed with scikit-learn

# Email
sendgrid==6.11.0
aiosmtplib==3.0.1

# OpenAI
openai==1.10.0

# Utils
requests==2.31.0
beautifulsoup4==4.12.3
feedparser==6.0.11
praw==7.7.1

# Rate Limiting
slowapi==0.1.9

# CORS
fastapi-cors==0.0.6
